+++ libs/ui/src/components/molecules/inputWrapper/fieldWrapper.tsx
@@ -1,14 +1,31 @@
 import classNames from 'classnames';
 import React from 'react';
 import { Field } from '../../../types/formAndField';
-import { HelperTextFactory } from '../../atoms/helperTexts';
 import InputLabel from '../../atoms/inputLabel/inputLabel';
 import { FieldWrapperBase } from './fieldWrapperStyles';
+import {
+  ErrorMessage,
+  WarningMessage,
+  SuccessMessage,
+} from '../../../types/formAndField';
+import {
+  ErrorText,
+  HelperText,
+  SuccessText,
+  WarningText,
+} from '../../atoms/helperTexts';
 
 type FieldWrapperProps = Field & {
   children: React.ReactNode;
   long?: boolean;
   full?: boolean;
+};
+
+export type HelperTextFactoryProps = {
+  error?: ErrorMessage;
+  warning?: WarningMessage;
+  success?: SuccessMessage;
+  helperText?: string;
 };
 
 const FieldWrapper = ({
@@ -42,3 +59,17 @@ const FieldWrapper = ({
 };
 
 export default FieldWrapper;
+
+const HelperTextFactory = ({
+  error,
+  success,
+  warning,
+  helperText,
+}: HelperTextFactoryProps) => {
+  if (!error && !success && !warning && !helperText) return null;
+
+  if (error) return <ErrorText>{error.message}</ErrorText>;
+  if (warning) return <WarningText>{warning.message}</WarningText>;
+  if (success) return <SuccessText>{success.message}</SuccessText>;
+  return <HelperText>{helperText}</HelperText>;
+};
